optimize interactive filtering
Before this commit, if the source was huge, whenever we typed a character,
Vim needed too much time to filter the source.

Not anymore.  Now, we split the source in chunks.  The first one is
processed immediately; the next ones are processed with timers to let us
continue typing characters if needed.